Fixes #2953: Add info.txt with debug info to Database Export
