+++ podcasts/DatabaseExport.swift
@@ -70,6 +70,9 @@ class DatabaseExport {
             let logsFile = exportDirectory.appendingPathComponent("logs.txt", isDirectory: false)
             try fileManager.copyItem(at: URL(fileURLWithPath: logFile), to: logsFile)
 
+            let debugInfoFile = exportDirectory.appendingPathComponent("info.txt", conformingTo: .text)
+            fileManager.createFile(atPath: debugInfoFile.absoluteString, contents: DebugInfo.string(optOut: UserDefaults.standard.debugOptedOut).data(using: .utf8))
+
             // Write the bundle document
             let exportFile = exportDirectory.appendingPathComponent("export", conformingTo: .pcasts)
             let wrapper = try PCBundleDoc().fileWrapper()
